app/backend/requirements.txt
requests
mongoengine
flask
flask-restful
flask_pymongo
dnspython
passlib
pyjwt
flask_cors
